Move handle to a more logical location and impl FromStr
