+++ src/structs/types/memory_array_mapped_address.rs
@@ -1,4 +1,5 @@
-use crate::{Handle, SMBiosStruct, UndefinedStruct};
+use crate::core::{Handle, UndefinedStruct};
+use crate::SMBiosStruct;
 use std::fmt;
 
 /// # Memory Array Mapped Address (Type 19)
