+++ src/structs/types/processor_information.rs
@@ -1,4 +1,5 @@
-use crate::{Handle, SMBiosStruct, UndefinedStruct};
+use crate::core::{Handle, UndefinedStruct};
+use crate::SMBiosStruct;
 use std::convert::TryInto;
 use std::fmt;
 use std::ops::Deref;
